refactor: remove trpc/react imports from packages/features
- Create server-safe RouterOutputs/RouterInputs export in @calcom/trpc/server/router-types
- Update packages/features/bookings/types.ts to import from server-safe location
- Update packages/features/eventtypes/lib/types.ts to import from server-safe location
- Move useApp.ts hook from packages/features to packages/app-store/_utils
- Add biome rule to prevent packages/features from importing @calcom/trpc/react
