+++ packages/features/bookings/types.ts
@@ -3,7 +3,7 @@ import type { ErrorOption, FieldPath } from "react-hook-form";
 import type { RegularBookingCreateResult } from "@calcom/features/bookings/lib/dto/types";
 import type { Slots } from "@calcom/features/calendars/lib/types";
 import type { SchedulingType } from "@calcom/prisma/enums";
-import type { RouterOutputs } from "@calcom/trpc/react";
+import type { RouterOutputs } from "@calcom/trpc/server/router-types";
 import type { AppsStatus } from "@calcom/types/Calendar";
 
 import type { BookingCreateBody } from "./lib/bookingCreateBodySchema";
